Feat: add german language support (#9892)
**What**
Added German language json file. 

**Why**
To be able to choose german language.

**How**
The File is translated with Google Translate, some human checks and corrections. I'm sure it's not perfect, but it is a good starting point. First the en.json was duplicated (as de.json) to be able to see the changes made and if the structure is still the same.

**Testing**
As mentioned in How, i checked if the structure is still the same.

Co-authored-by: Kasper Fabricius Kristensen <45367945+kasperkristensen@users.noreply.github.com>